Dropped all of the sub-testX scripts.
The doctest-tools testall script will work from the root source directory now
for both Python 2.5 and 2.6.  But it doesn't delete all of the compiled_krb
files first.

So we're not quite done upgrading the testing yet!  But there are a lot of
changes to this point, and I wanted to checkpoint everything...